feat: refactor FirstFit memory management class (#master)
This commit refactors the FirstFit memory management class. The changes
include eliminating redundant list objects, simplifying the alloc and release
methods, and improving the usage of the MemoryBlock class. The commit also
includes the implementation of a method to merge adjacent free blocks
and makes the printLayout method more readable by separating allocated and
free blocks. This method is not functioning as expected and its usage is therefore commented out for now.

Furthermore, the commit involves changes to the MemoryBlock
class to include an 'allocated' attribute for easier management of memory blocks.